data_stitching/requirements.txt
numpy==1.23.5
Pillow==9.4.0
Pillow==10.0.1
